Unoptimized version
I cobbled the win condition check function in a few minutes. Turned out that the bug I was experiencing was actually an unsigned int being turned into a signed int and that was messing up the win check conditions. I have a few optimizations in mind already to speed up the program or at least reduce line numbers and complexity.